_metadata_from_built_wheel: pass metadata_directory=None
Instruct the backend to generate the metadata when building the wheel.

After build_wheel returns, `_metadata_from_built_wheel` extracts the metadata
into the passed `metadata_directory` so subsequent calls to build_wheel (if
made), should have a good copy of the metadata in the place they expect to find
it.

Fix #47