docs: set html_last_updated_fmt to format string
According to the Sphinx documentation and source code,
html_last_updated_fmt is supposed to be a strftime()-like format string,
not a literal string with a date.

Instead of datetime.now().isoformat, set this to the equivalent format
string (sans microseconds, as Sphinx does not support %f).

This is a no-op generally. However, unlike our implementation here,
Sphinx's date generation obeys SOURCE_DATE_EPOCH, which means that the
builds will now be able to be built reproducibly.

Fixes #53

Signed-off-by: Faidon Liambotis <paravoid@debian.org>